// src/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      employees: {
        Row: {
          created_at: string | null
          email: string | null
          id: number
          name: string | null
        }
        Insert: {
          created_at?: string | null
          email?: string | null
          id?: never
          name?: string | null
        }
        Update: {
          created_at?: string | null
          email?: string | null
          id?: never
          name?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}